To improve scalability I opted to turn the apis namespace package into a package.
This allows to have a central place for registering api namespaces to my
API blueprint. Currently this may not seem beneficial as there is
currently only one namespace registered. However, it allows to define
future namespaces to be defined in a seperate file and to be then
registered in a central file, the __init__.py file.

I tried to split the api routes in four different namespaces to seperate
things and make it easier to maintain. However, the interdependent
nature of the api models involved lead to circular imports.

After thinking about it a little more, I reckong it is not a bad idea
after all to define api routes concerning data on books in one central
place.